feat: New Testing UI in dev mode, separated examples and tests
