+++ scripts/constants.js
@@ -1,0 +1,3 @@
+export const DIST_DIR = "dist";
+export const SRC_DIR = "src";
+export const SRC_PATH = `${SRC_DIR}/index.ts`;
