+++ scripts/lib/autoinput.js
